+++ Assets/MiscAssets/PolygonHeist/Model/SM_Wep_RifleSwat_01.fbx.meta
@@ -1,95 +1,201 @@
 fileFormatVersion: 2
 guid: b2a22edd3225cae488adbefe555f5e81
-timeCreated: 1502255135
-licenseType: Store
 ModelImporter:
-  serializedVersion: 18
-  fileIDToRecycleName:
-    100000: //RootNode
-    100002: SM_Wep_RifleSwat_ChargeHandle_01
-    100004: SM_Wep_RifleSwat_Mag_01
-    100006: SM_Wep_RifleSwat_Slide_01
-    100008: SM_Wep_RifleSwat_Stock_01
-    100010: SM_Wep_RifleSwat_Trigger_01
-    400000: //RootNode
-    400002: SM_Wep_RifleSwat_ChargeHandle_01
-    400004: SM_Wep_RifleSwat_Mag_01
-    400006: SM_Wep_RifleSwat_Slide_01
-    400008: SM_Wep_RifleSwat_Stock_01
-    400010: SM_Wep_RifleSwat_Trigger_01
-    2300000: //RootNode
-    2300002: SM_Wep_RifleSwat_ChargeHandle_01
-    2300004: SM_Wep_RifleSwat_Mag_01
-    2300006: SM_Wep_RifleSwat_Slide_01
-    2300008: SM_Wep_RifleSwat_Stock_01
-    2300010: SM_Wep_RifleSwat_Trigger_01
-    3300000: //RootNode
-    3300002: SM_Wep_RifleSwat_ChargeHandle_01
-    3300004: SM_Wep_RifleSwat_Mag_01
-    3300006: SM_Wep_RifleSwat_Slide_01
-    3300008: SM_Wep_RifleSwat_Stock_01
-    3300010: SM_Wep_RifleSwat_Trigger_01
-    4300000: SM_Wep_RifleSwat_01
-    4300002: SM_Wep_RifleSwat_Mag_01
-    4300004: SM_Wep_RifleSwat_Stock_01
-    4300006: SM_Wep_RifleSwat_Slide_01
-    4300008: SM_Wep_RifleSwat_Trigger_01
-    4300010: SM_Wep_RifleSwat_ChargeHandle_01
-    9500000: //RootNode
+  serializedVersion: 22200
+  internalIDToNameTable:
+  - first:
+      1: 100000
+    second: //RootNode
+  - first:
+      1: 100002
+    second: SM_Wep_RifleSwat_ChargeHandle_01
+  - first:
+      1: 100004
+    second: SM_Wep_RifleSwat_Mag_01
+  - first:
+      1: 100006
+    second: SM_Wep_RifleSwat_Slide_01
+  - first:
+      1: 100008
+    second: SM_Wep_RifleSwat_Stock_01
+  - first:
+      1: 100010
+    second: SM_Wep_RifleSwat_Trigger_01
+  - first:
+      4: 400000
+    second: //RootNode
+  - first:
+      4: 400002
+    second: SM_Wep_RifleSwat_ChargeHandle_01
+  - first:
+      4: 400004
+    second: SM_Wep_RifleSwat_Mag_01
+  - first:
+      4: 400006
+    second: SM_Wep_RifleSwat_Slide_01
+  - first:
+      4: 400008
+    second: SM_Wep_RifleSwat_Stock_01
+  - first:
+      4: 400010
+    second: SM_Wep_RifleSwat_Trigger_01
+  - first:
+      23: 2300000
+    second: //RootNode
+  - first:
+      23: 2300002
+    second: SM_Wep_RifleSwat_ChargeHandle_01
+  - first:
+      23: 2300004
+    second: SM_Wep_RifleSwat_Mag_01
+  - first:
+      23: 2300006
+    second: SM_Wep_RifleSwat_Slide_01
+  - first:
+      23: 2300008
+    second: SM_Wep_RifleSwat_Stock_01
+  - first:
+      23: 2300010
+    second: SM_Wep_RifleSwat_Trigger_01
+  - first:
+      33: 3300000
+    second: //RootNode
+  - first:
+      33: 3300002
+    second: SM_Wep_RifleSwat_ChargeHandle_01
+  - first:
+      33: 3300004
+    second: SM_Wep_RifleSwat_Mag_01
+  - first:
+      33: 3300006
+    second: SM_Wep_RifleSwat_Slide_01
+  - first:
+      33: 3300008
+    second: SM_Wep_RifleSwat_Stock_01
+  - first:
+      33: 3300010
+    second: SM_Wep_RifleSwat_Trigger_01
+  - first:
+      43: 4300000
+    second: SM_Wep_RifleSwat_01
+  - first:
+      43: 4300002
+    second: SM_Wep_RifleSwat_Mag_01
+  - first:
+      43: 4300004
+    second: SM_Wep_RifleSwat_Stock_01
+  - first:
+      43: 4300006
+    second: SM_Wep_RifleSwat_Slide_01
+  - first:
+      43: 4300008
+    second: SM_Wep_RifleSwat_Trigger_01
+  - first:
+      43: 4300010
+    second: SM_Wep_RifleSwat_ChargeHandle_01
+  - first:
+      95: 9500000
+    second: //RootNode
+  externalObjects: {}
   materials:
-    importMaterials: 0
+    materialImportMode: 0
     materialName: 0
     materialSearch: 1
+    materialLocation: 0
   animations:
     legacyGenerateAnimations: 4
     bakeSimulation: 0
+    resampleCurves: 1
     optimizeGameObjects: 0
+    removeConstantScaleCurves: 0
     motionNodeName: 
+    rigImportErrors: 
+    rigImportWarnings: 
+    animationImportErrors: 
+    animationImportWarnings: 
+    animationRetargetingWarnings: 
+    animationDoRetargetingWarnings: 0
+    importAnimatedCustomProperties: 0
+    importConstraints: 0
     animationCompression: 1
-    animationRotationError: .5
-    animationPositionError: .5
-    animationScaleError: .5
+    animationRotationError: 0.5
+    animationPositionError: 0.5
+    animationScaleError: 0.5
     animationWrapMode: 0
     extraExposedTransformPaths: []
+    extraUserProperties: []
     clipAnimations: []
-    isReadable: 0
+    isReadable: 1
   meshes:
     lODScreenPercentages: []
     globalScale: 1
     meshCompression: 0
     addColliders: 0
+    useSRGBMaterialColor: 1
+    sortHierarchyByName: 1
+    importPhysicalCameras: 0
+    importVisibility: 0
     importBlendShapes: 0
+    importCameras: 0
+    importLights: 0
+    nodeNameCollisionStrategy: 0
+    fileIdsGeneration: 1
     swapUVChannels: 0
     generateSecondaryUV: 0
     useFileUnits: 1
-    optimizeMeshForGPU: 1
     keepQuads: 0
     weldVertices: 1
+    bakeAxisConversion: 0
+    preserveHierarchy: 0
+    skinWeightsMode: 0
+    maxBonesPerVertex: 4
+    minBoneWeight: 0.001
+    optimizeBones: 1
+    meshOptimizationFlags: -1
+    indexFormat: 1
     secondaryUVAngleDistortion: 8
     secondaryUVAreaDistortion: 15.000001
     secondaryUVHardAngle: 88
+    secondaryUVMarginMethod: 0
+    secondaryUVMinLightmapResolution: 40
+    secondaryUVMinObjectScale: 1
     secondaryUVPackMargin: 4
     useFileScale: 1
+    strictVertexDataChecks: 0
   tangentSpace:
     normalSmoothAngle: 5
-    splitTangentsAcrossUV: 0
     normalImportMode: 1
     tangentImportMode: 1
+    normalCalculationMode: 0
+    legacyComputeAllNormalsFromSmoothingGroupsWhenMeshHasBlendShapes: 1
+    blendShapeNormalImportMode: 1
+    normalSmoothingSource: 0
+  referencedClips: []
   importAnimation: 0
-  copyAvatar: 0
   humanDescription:
+    serializedVersion: 3
     human: []
     skeleton: []
-    armTwist: .5
-    foreArmTwist: .5
-    upperLegTwist: .5
-    legTwist: .5
-    armStretch: .0500000007
-    legStretch: .0500000007
+    armTwist: 0.5
+    foreArmTwist: 0.5
+    upperLegTwist: 0.5
+    legTwist: 0.5
+    armStretch: 0.05
+    legStretch: 0.05
     feetSpacing: 0
+    globalScale: 1
     rootMotionBoneName: 
+    hasTranslationDoF: 0
+    hasExtraRoot: 0
+    skeletonHasParents: 0
   lastHumanDescriptionAvatarSource: {instanceID: 0}
+  autoGenerateAvatarMappingIfUnspecified: 1
   animationType: 0
+  humanoidOversampling: 1
+  avatarSetup: 0
+  addHumanoidExtraRootOnlyWhenUsingAvatar: 0
+  importBlendShapeDeformPercent: 0
+  remapMaterialsIfMaterialImportModeIsNone: 1
   additionalBone: 0
   userData: 
   assetBundleName:
