+++ Assets/_MainAssets/Prefabs/Valuables/Valuable_Figure_02.prefab
@@ -120,6 +120,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   scoreValue: 50
-  isChecklistItem: 0
-  type: 3
+  type: 2
   scoreScript: {fileID: 0}
+  checklistController: {fileID: 0}
