+++ Assets/_MainAssets/Prefabs/Valuables/Valuable_Trophy_01.prefab
@@ -171,6 +171,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   scoreValue: 50
-  isChecklistItem: 0
-  type: 2
+  type: 1
   scoreScript: {fileID: 0}
+  checklistController: {fileID: 0}
