+++ Assets/_MainAssets/Scripts/DevTools/TimerManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 78687633ab7aab042ad3ebdcf5c931f4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
