+++ Assets/_MainAssets/Scripts/Interactables/Props/EscapeController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 35c02369c7e3ad644a7e0e4ddfc1a2e7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
