+++ Assets/_MainAssets/Scripts/Interactables/Props/EscapeVan.cs
@@ -1,0 +1,19 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+public class EscapeVan : MonoBehaviour, IInteractable
+{
+    public void Interact()
+    {
+        EscapeController escapeController = GetComponent<EscapeController>();
+        if (escapeController.canEscape && !escapeController.timesUp)
+        {
+            escapeController.EscapeSuccess();
+        }
+
+    }
+
+}
+
