+++ Assets/_MainAssets/Scripts/Interactables/Props/EscapeVan.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cd3754bd678c1b64499eab436160daf9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
