--- Assets/_MainAssets/Scripts/Player/UI.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7c909cdd354744747ae980de89a24ad7
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
