--- Assets/_MainAssets/Scripts/Player/UI/Checklist.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4b4f6f5c6a173db4a9d376dcfe3bb0f9
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
