--- Assets/_MainAssets/Scripts/Player/UI/Checklist/ChecklistController.cs
@@ -1,24 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class ChecklistController : MonoBehaviour
-{
-    [Header("Checklist Items")]
-    [SerializeField]
-    private ChecklistItem[] checklistItems;
-    
-    // Start is called before the first frame update
-    void Start()
-    {
-          
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-
-    
-}
