--- Assets/_MainAssets/Scripts/Player/UI/Checklist/ChecklistController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5a775f641309e9f4eaf9515f456dde2b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
