--- Assets/_MainAssets/Scripts/Player/UI/StartButton.cs
@@ -1,13 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class StartButton : MonoBehaviour
-{
-    public void HandleStartButton()
-    {
-        SceneManager.LoadScene("GameScene");
-    }
-
-}
