--- Assets/_MainAssets/Scripts/Player/UI/StartButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b8e33f8ab2d91445bb8fa6fd5fe44cf3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
