--- Assets/_MainAssets/Scripts/Player/UI/StatusDisplay.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 930eb9f1e77d67c4c9eee0a3b2e30f77
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
