--- Assets/_MainAssets/Scripts/Player/UI/WheelMenu.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 20eee3eac5151ce48a202b5f9a804dc9
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
