--- Assets/_MainAssets/Scripts/Player/UI/WheelMenu/MenuDisplayScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6a0f96a5075e5ea49be75daba08ef4ac
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
