--- Assets/_MainAssets/Scripts/Player/UI/WheelMenu/WheelMenuScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 880e34271eefe6543b94494cdd1c20f9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
