+++ Assets/_MainAssets/Scripts/UI/Checklist.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4b4f6f5c6a173db4a9d376dcfe3bb0f9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
