+++ Assets/_MainAssets/Scripts/UI/Checklist/ChecklistItem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b72316ed2c824d8418a5810a81a6b2d5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
