+++ Assets/_MainAssets/Scripts/UI/Score.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bbe285a898c7d419fadba08a1e629823
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
