+++ Assets/_MainAssets/Scripts/UI/StartButton.cs
@@ -1,0 +1,13 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+public class StartButton : MonoBehaviour
+{
+    public void HandleStartButton()
+    {
+        SceneManager.LoadScene("GameScene");
+    }
+
+}
