+++ Assets/_MainAssets/Scripts/UI/WheelMenu/MenuItemScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aabbc588a25ff2c47b2df9fe3907c9cc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
